requirements-celery.txt
amqp==5.2.0
billiard==3.6.4.0
celery==5.0.5
future==1.0.0
kombu==5.3.7
pytz==2024.1
redis==5.0.3
vine==5.1.0
loguru==0.7.2
python-dotenv==1.0.1
transformers==4.39.3
pillow==10.3.0
sentence-transformers==2.6.1
openai==1.16.2
typing==3.7.4.3
nltk==3.8.1
pandas==2.2.1
simplejson==3.19.2
